Move cohort selection counts to counts service.
